requirements.txt
streamlit==1.32.2
opencv-python==4.9.0.80
streamlit-sortables==0.2.0
numpy==1.26.4
scikit-image==0.22.0
Pillow==10.3.0
